feat: finalizado!

+++ src/controllers/projectController.js
@@ -6,11 +6,25 @@ const User = require('../models/User');
 
 router.use(authMiddleware);
 
-router.get('/', async (req, res) => {
+router.get('/usuarios', async (req, res) => {
   try {
-    const users = await User.find();
+    const users = await User.findById(req.userId);
+    console.log(users)
+    // if (users.level != 999) {
+    //   return res.status(401).send({ error: 'Usuário não autorizado' });
+    // }
+    const usuarios = await User.find({
+      $or: [
+        {
+          level: 0,
+        },
+        {
+          level: 1,
+        },
+      ],
+    });
 
-    return res.send({ users });
+    return res.send({ usuarios });
   } catch (err) {
     return res.status(400).send({ error: 'Erro ao listar todos os usuarios' });
   }
@@ -24,7 +38,7 @@ router.get('/:userId', async (req, res) => {
     return res.status(400).send({ error: 'Erro ao listar o usuario' });
   }
 });
-router.put('/update/:idUser', async function (req, res) {
+router.put('/update/:userId', async function (req, res) {
 
   try{
   const newUser = req.body;
@@ -32,23 +46,64 @@ router.put('/update/:idUser', async function (req, res) {
   if(newUser.password) {
     newUser.password = await bcrypt.hashSync(req.body.password, 10);
   }
-  console.log(newUser);
-  const user = await User.findByIdAndUpdate(req.params.idUser, newUser, {new: true,})
+  const usuario = await User.findByIdAndUpdate(req.params.userId, newUser, {
+    new: true,
+  });
 
+  await usuario.save();
+
+  return res.send({ usuario })
   
-  
-  await user.save()
-  return res.send({ user })
 }catch (err) {
   console.log(err);
   return res.status(400).send({ error: "Erro no alterar"});
 }
 
 });
+
+
+router.put('/update/DesativarUsuario/:userId', async function (req, res) {
+  try {
+    const newData = req.body;
+    //atualizando e retornando usuario atualizado
+    if (newData.password)
+      newData.password = await bcrypt.hashSync(req.body.password, 10);
+    const usuario = await User.findByIdAndUpdate(req.params.idUser, {
+      level: 0,
+    });
+    await usuario.save();
+    return res.send({ usuario });
+  } catch (err) {
+    console.log(err);
+    return res.send({ error: 'Erro ao atualizar usuário!' });
+  }
+});
+
+router.put('/update/AtivarUsuario/:userId', async function (req, res) {
+  try {
+    const newData = req.body;
+    //atualizando e retornando usuario atualizado
+    if (newData.password)
+      newData.password = await bcrypt.hashSync(req.body.password, 10);
+    const usuario = await User.findByIdAndUpdate(req.params.idUser, {
+      level: 1,
+    });
+    await usuario.save();
+    return res.send({ usuario });
+  } catch (err) {
+    console.log(err);
+    return res.send({ error: 'Erro ao atualizar usuário!' });
+  }
+});
+
 router.delete('/:userId', async (req, res) => {
   try {
+    const useradm = await User.findById(req.userId);
+    if (useradm.level !== 999) {
+      return res.send({ error: 'Sem autorização de exclusão' });
+    }
+    
     await User.findByIdAndRemove(req.params.userId);
-
     return res.send();
   } catch (err) {
     return res.status(400).send({ error: 'Erro ao deletar o usuario' });
